+++ selfdrive/car/volkswagen/interface.py
@@ -104,21 +104,11 @@ class CarInterface(CarInterfaceBase):
 
     events = self.create_common_events(ret, extra_gears=[GEAR.eco, GEAR.sport])
 
-    # Vehicle operation safety checks and events
+    # Vehicle health and operation safety checks
     if self.CS.parkingBrakeSet:
       events.append(create_event('parkBrake', [ET.NO_ENTRY, ET.USER_DISABLE]))
-
-    # Vehicle health safety checks and events
-    if self.CS.accFault:
-      events.append(create_event('radarFault', [ET.NO_ENTRY, ET.IMMEDIATE_DISABLE]))
     if self.CS.steeringFault:
       events.append(create_event('steerTempUnavailable', [ET.NO_ENTRY, ET.WARNING]))
-
-    # Per the Comma safety model, disable on pedals rising edge or when brake
-    # is pressed and speed isn't zero.
-    if (ret.gasPressed and not self.gas_pressed_prev) or \
-            (ret.brakePressed and (not self.brake_pressed_prev or not ret.standstill)):
-      events.append(create_event('pedalPressed', [ET.NO_ENTRY, ET.USER_DISABLE]))
 
     # Engagement and longitudinal control using stock ACC. Make sure OP is
     # disengaged if stock ACC is disengaged.
@@ -140,7 +130,7 @@ class CarInterface(CarInterfaceBase):
     self.buttonStatesPrev = self.CS.buttonStates.copy()
 
     self.CS.out = ret.as_reader()
-    return  self.CS.out
+    return self.CS.out
 
   def apply(self, c):
     can_sends = self.CC.update(c.enabled, self.CS, self.frame, c.actuators,