+++ selfdrive/logcatd/logcatd.cc
@@ -25,8 +25,8 @@ int main() {
   assert(system_logger);
   struct logger *crash_logger = android_logger_open(logger_list, LOG_ID_CRASH);
   assert(crash_logger);
-//  struct logger *kernel_logger = android_logger_open(logger_list, LOG_ID_KERNEL);
-//  assert(kernel_logger);
+  struct logger *kernel_logger = android_logger_open(logger_list, (log_id_t)5); // LOG_ID_KERNEL
+  assert(kernel_logger);
 
   Context * c = Context::create();
   PubSocket * androidLog = PubSocket::create(c, "androidLog");
@@ -48,7 +48,7 @@ int main() {
     capnp::MallocMessageBuilder msg;
     cereal::Event::Builder event = msg.initRoot<cereal::Event>();
     event.setLogMonoTime(nanos_since_boot());
-    auto androidEntry = event.initAndroidLogEntry();
+    auto androidEntry = event.initAndroidLog();
     androidEntry.setId(log_msg.id());
     androidEntry.setTs(entry.tv_sec * 1000000000ULL + entry.tv_nsec);
     androidEntry.setPriority(entry.priority);