+++ selfdrive/loggerd/ethernetsniffer.py
@@ -8,8 +8,7 @@ def main():
   ethernetData = messaging.pub_sock('ethernetData')
 
   for ts, pkt in pcap.pcap('eth0'):
-    dat = messaging.new_message()
-    dat.init('ethernetData', 1)
+    dat = messaging.new_message('ethernetData', 1)
     dat.ethernetData[0].ts = ts
     dat.ethernetData[0].pkt = str(pkt)
     ethernetData.send(dat.to_bytes())