+++ selfdrive/modeld/models/driving.h
@@ -34,6 +34,7 @@
 #define MODEL_PATH_DISTANCE 192
 #define POLYFIT_DEGREE 4
 #define SPEED_PERCENTILES 10
+#define DESIRE_LEN 8
 #define DESIRE_PRED_SIZE 32
 #define OTHER_META_SIZE 4
 #define LEAD_MDN_N 5 // probs for 5 groups
@@ -51,6 +52,7 @@ struct ModelDataRaw {
     float *long_x;
     float *long_v;
     float *long_a;
+    float *desire_state;
     float *meta;
     float *pose;
   };
@@ -63,6 +65,7 @@ typedef struct ModelState {
   RunModel *m;
 #ifdef DESIRE
   float *desire;
+  float *pulse_desire;
 #endif
 } ModelState;