+++ selfdrive/modeld/test/opencl_hooks/build.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+gcc -fPIC -I /data/openpilot/phonelibs/opencl/include -shared hook.c
+